+++ src/sewerrat/__init__.py
@@ -24,3 +24,5 @@ from .retrieve_directory import retrieve_directory
 from .retrieve_file import retrieve_file
 from .retrieve_metadata import retrieve_metadata
 from .list_registered_directories import list_registered_directories
+from .list_tokens import list_tokens
+from .list_fields import list_fields