+++ src/sewerrat/list_registered_directories.py
@@ -1,6 +1,8 @@
-from typing import Optional, Union, List, Dict
+from typing import Optional, Union, List, Dict, Literal
 import requests
 import urllib
+
+from . import _utils as ut
 
 
 def list_registered_directories(
@@ -10,6 +12,8 @@ def list_registered_directories(
     within: Optional[str] = None,
     prefix: Optional[str] = None,
     exists: Optional[bool] = None,
+    number: int = 100, 
+    on_truncation: Literal["message", "warning", "none"] = "message"
 ) -> List[Dict]:
     """
     List all registered directories in the SewerRat instance.
@@ -19,9 +23,9 @@ def list_registered_directories(
             URL to the SewerRat REST API.
 
         user:
-            Name of a user. If not None, this is used to filter the returned
-            directories based on the user who registered them. Alternatively
-            True, to automatically use the name of the current user.
+            Name of a user.
+            If not ``None``, this is used to filter the returned directories based on the user who registered them.
+            Alternatively, this can be set to ``True`` to automatically use the name of the current user.
 
         contains:
             String containing an absolute path. If not None, results are
@@ -40,11 +44,20 @@ def list_registered_directories(
             Whether to only report directories that exist on the filesystem.
             If ``False``, only non-existent directories are reported, and if ``None``, no filtering is applied based on existence.
 
+        number:
+            Integer specifying the maximum number of results to return.
+            This can also be ``float("inf")`` to retrieve all available results.
+
+        on_truncation:
+            String specifying the action to take when the number of search results is capped by ``number``.
+
     Returns:
-        List of objects where each object corresponds to a registered directory
-        and contains the `path` to the directory, the `user` who registered it,
-        the Unix epoch `time` of the registration, and the `names` of the
-        metadata files to be indexed.
+        List of dictionaries where each dictionary corresponds to a registered directory and contains:
+
+        - `path`, the path to the directory.
+        - `user`, the name of the user who registered it.
+        - `time`, the Unix epoch `time` of the registration.
+        - `names`, a list containing the names of the metadata files to be indexed.
     """
     query = []
     if not user is None and user != False:
@@ -65,11 +78,34 @@ def list_registered_directories(
             qstr = "false"
         query.append("exists=" + qstr)
 
-    url += "/registered"
+    stub = "/registered"
+    use_question = True
     if len(query) > 0:
-        url += "?" + "&".join(query)
+        stub += "?" + "&".join(query)
+        use_question = False
 
-    res = requests.get(url)
-    if res.status_code >= 300:
-        raise ut.format_error(res)
-    return res.json()
+    original_number = number
+    if on_truncation != "none":
+        number += 1
+
+    collected = []
+    while len(collected) < number:
+        current_url = url + stub
+        if number != float("inf"):
+            sep = "&"
+            if use_question:
+                sep = "?"
+            current_url += sep + "limit=" + str(number - len(collected))
+
+        res = requests.get(current_url)
+        if res.status_code >= 300:
+            raise ut.format_error(res)
+        payload = res.json()
+        collected += payload["results"]
+
+        if "next" not in payload:
+            break
+        stub = payload["next"]
+        use_question = False
+
+    return ut.handle_truncated_pages(on_truncation, original_number, collected)